refactor(console): polish ui
