+++ packages/console/src/hooks/use-console-routes/routes/mfa.tsx
@@ -1,0 +1,5 @@
+import { type RouteObject } from 'react-router-dom';
+
+import Mfa from '@/pages/Mfa';
+
+export const mfa: RouteObject = { path: 'mfa', element: <Mfa /> };
