+++ packages/console/src/pages/ApplicationDetails/ApplicationDetailsContent/index.tsx
@@ -15,6 +15,7 @@ import ApplicationIcon from '@/components/ApplicationIcon';
 import DetailsForm from '@/components/DetailsForm';
 import DetailsPageHeader from '@/components/DetailsPage/DetailsPageHeader';
 import Drawer from '@/components/Drawer';
+import EmptyDataPlaceholder from '@/components/EmptyDataPlaceholder';
 import OrganizationList from '@/components/OrganizationList';
 import UnsavedChangesAlertModal from '@/components/UnsavedChangesAlertModal';
 import { ApplicationDetailsTabs, logtoThirdPartyGuideLink, protectedAppLink } from '@/consts';
@@ -22,7 +23,9 @@ import { isDevFeaturesEnabled } from '@/consts/env';
 import DeleteConfirmModal from '@/ds-components/DeleteConfirmModal';
 import TabNav, { TabNavItem } from '@/ds-components/TabNav';
 import TabWrapper from '@/ds-components/TabWrapper';
+import TextLink from '@/ds-components/TextLink';
 import useApi from '@/hooks/use-api';
+import { organizations } from '@/hooks/use-console-routes/routes/organizations';
 import useDocumentationUrl from '@/hooks/use-documentation-url';
 import useTenantPathname from '@/hooks/use-tenant-pathname';
 import { applicationTypeI18nKey } from '@/types/applications';
@@ -237,7 +240,20 @@ function ApplicationDetailsContent({ data, oidcConfig, onApplicationUpdated }: P
             isActive={tab === ApplicationDetailsTabs.Organizations}
             className={styles.tabContainer}
           >
-            <OrganizationList type="application" data={data} />
+            <OrganizationList
+              type="application"
+              data={data}
+              placeholder={
+                <EmptyDataPlaceholder
+                  title={
+                    <Trans
+                      i18nKey="admin_console.application_details.no_organization_placeholder"
+                      components={{ a: <TextLink to={'/' + organizations.path} /> }}
+                    />
+                  }
+                />
+              }
+            />
           </TabWrapper>
         </>
       )}
